+++ packages/dara-core/changelog.md
@@ -2,11 +2,13 @@
 title: Changelog
 ---
 
-## NEXT
+## 1.7.4
 
 -   Added the ability to pass an asynchronous function to `ConfigurationBuilder.on_startup(...)`.
 -   Fix an issue where `get_settings` would crash when attempting to generate a missing `.env` file. It now
 instead warns and falls back to using an in-memory set of default settings.
+-   Fix an issue where the `BackendStore` would not respect existing value and overwrite it with the variable default,
+e.g. when an existing JSON file was present on disk with the `FileBackend`
 
 ## 1.7.3